+++ pages/search/result/[name].tsx
@@ -2,10 +2,11 @@ import Head from "next/head";
 import { useState } from "react";
 import { useRouter } from "next/router";
 import SearchBoxComponent from "../../../src/components/search/common/SearchBoxComponent";
-import CategoryComponent from "../../../src/components/search/result/CategoryComponent";
+import FilterCategoryComponent from "../../../src/components/search/result/FilterCategoryComponent";
+import FilterResultComponent from "../../../src/components/search/result/FilterResultComponent";
 import SearchInitComponent from "../../../src/components/search/result/SearchInitComponent";
 import NotFoundComponent from "../../../src/components/search/result/NotFoundComponent";
-import SearchResultComponent from "../../../src/components/search/result/SearchResultComponent";
+import FilterKindComponent from "../../../src/components/search/result/FilterKindComponent";
 import ListBookComponent from "src/components/search/result/ListBookComponent";
 import ListStudyComponent from "src/components/search/result/ListStudyComponent";
 
@@ -19,13 +20,14 @@ export default function Detail(): JSX.Element {
         <title>검색 결과 : {Router.query.name}</title>
       </Head>
       <SearchBoxComponent />
-      <CategoryComponent />
-      <SearchInitComponent />
-      <NotFoundComponent />
-      <SearchResultComponent
+      <FilterKindComponent
         tabSelected={tabSelected}
         setTabSelected={setTabSelected}
       />
+      <FilterCategoryComponent />
+      <FilterResultComponent />
+      <SearchInitComponent />
+      <NotFoundComponent />
       {[0, 1].includes(tabSelected) && <ListBookComponent />}
       {[0, 2].includes(tabSelected) && <ListStudyComponent />}
     </>
