+++ src/components/search/result/NotFoundComponent.tsx
@@ -22,7 +22,7 @@ const NotFoundSub = styled.div`
   position: relative;
   padding: 10px 0 109px 0;
   line-height: 20px;
-  color: #222222;
+  color: #999999;
   font-size: 14px;
 `;
 
@@ -35,7 +35,7 @@ export default function NotFoundComponent(): JSX.Element {
       <NotFoundMain>검색 결과가 없습니다.</NotFoundMain>
       <NotFoundSub>
         다른 검색어를 입력하시거나
-        {"\n"}
+        <br />
         철자와 띄어쓰기를 확인해보세요.
       </NotFoundSub>
     </NotFoundWrapper>
