+++ src/components/search/result/SearchInitComponent.tsx
@@ -10,7 +10,7 @@ export default function SearchInitComponent(): JSX.Element {
   return (
     <SearchInit>
       책, 저자, 스터디명을 검색하며 찾아보실 수 있어요.
-      {"\n"}
+      <br />
       철자와 띄어쓰기를 정확하게 입력해주세요.
     </SearchInit>
   );
